refactor: Use the stop propagation fn as the prevent default
