+++ src/components/CollectionDetailPage/CollectionItem/CollectionItem.tsx
@@ -5,7 +5,7 @@ import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { Dropdown, Icon, Button, Mana, Table, Popup } from 'decentraland-ui'
 import { Link } from 'react-router-dom'
 import { locations } from 'routing/locations'
-import { preventDefault } from 'lib/preventDefault'
+import { preventDefault } from 'lib/event'
 import { isComplete, isFree, canManageItem, getMaxSupply } from 'modules/item/utils'
 import ItemStatus from 'components/ItemStatus'
 import { isLocked } from 'modules/collection/utils'
