+++ src/components/FileImport/FileImport.tsx
@@ -1,15 +1,11 @@
 import * as React from 'react'
 import classNames from 'classnames'
 import Dropzone, { DropzoneState } from 'react-dropzone'
+import { stopPropagation } from 'lib/event'
 import { Props } from './FileImport.types'
 import './FileImport.css'
 
 export default class FileImport<T> extends React.Component<Props<T>> {
-  open = (openImporter: () => void) => (event: React.MouseEvent) => {
-    event.stopPropagation()
-    openImporter()
-  }
-
   renderDropZone = (props: DropzoneState) => {
     // TODO: Upgrade react-dropzone to a newer version to avoid the linting error: unbound-method
     // eslint-disable-next-line @typescript-eslint/unbound-method
@@ -27,7 +23,7 @@ export default class FileImport<T> extends React.Component<Props<T>> {
         {(!items || items.length === 0) && renderAction && (
           <span className="cta">
             <div className="image" />
-            <div>{renderAction(this.open(open))}</div>
+            <div>{renderAction(stopPropagation(open))}</div>
           </span>
         )}
       </div>
